get_all_names_done
