+++ dclnt.py
@@ -3,39 +3,23 @@ import os
 import collections
 from nltk import pos_tag, word_tokenize, punkt
 import constants
+import custom_methods
 
-# OPERATION AT LISTS
-def flattening(l):
-    if type(l) == 'List':
-        for el in l:
-            if isinstance(el, collections.Iterable) and not isinstance(el, basestring):
-                for sub in flattening(el):
-                    yield sub
-            else:
-                yield el
-    else:
-        return list(sum(_list,()))
+def flattening(array):
+    custom_methods.flattening(array)
 
-# OPERATION AT STRINGS
-def is_verb(word = None):
-    if word is None:
-        return False
-    else:
-        pos_info = pos_tag(word_tokenize(word))
-        return pos_info[0][1] in ('VB', 'VBD', 'VBZ', 'VBN')
-# FILTER FOR ARRAYS
+def is_verb(array = None):
+    custom_methods.is_verb(array)
+
 def is_none_filter(array):
-    node_list = [map(is_astFunction_instance_filter(y), ast.walk(y)) for y in array]
-    return filter(None, node_list) # OPTIMIZE <---------------------- need to delete somehow
-# FILTERING
+    custom_methods.is_none_filter(array)
+
 def filter_only_py_extention(file, from_path):
-    if file.endswith('.py'):
-        return os.path.join(from_path, file)
+    custom_methods.filter_only_py_extention(file, from_path)
 # SEARCHING
 def find_py_files(from_path = Path):
     files_list = []
     for whole_path, dirs, files in os.walk(from_path, topdown = True):
-        # files_list = [ filter_only_py_extention(file, whole_path) for file in files ] #if len(files_list) != 100 ]        
         for file in files:
             files_list.append(filter_only_py_extention(file, whole_path))
             if len(files_list) >= 100:
@@ -69,7 +53,7 @@ def get_verbs_from_function_name(function_name):
 
 def is_private_filter_and_stringify(thing):
     if True:
-    # FIX no clearly stated conditions 
+    # FIX no clearly understood conditions 
     # if thing.__class__.__name__.startswith('__') and thing.__class__.__name__.endswith('__'):
         # print thing.__class__.__name__
         return "%s" %thing.__class__.__name__
@@ -92,7 +76,7 @@ def cascade_call(path):
     py_files = find_py_files(path)
     trees = get_trees(py_files)
     return get_common_verbs(trees)
-
+# NOT FINISHED
 def get_common_verbs_across(path, projects):
     words = []
     for project in projects:
@@ -103,22 +87,17 @@ def get_common_verbs_across(path, projects):
         print(word, occurence)
 
 # UNUSED
-# 
-# def get_all_names(tree):
-#     return [node.id for node in ast.walk(tree) if isinstance(node, ast.Name)]
+def get_all_names(trees):
+    for t in trees:
+        names = [ node.id for node in ast.walk(t) if isinstance(node, ast.Name)]
+    return filter(None, names)
+
+def puts(smth):
+    print smth
 #
 # def split_snake_case_name_to_words(name):
 #     return [n for n in name.split('_') if n]
 #     return flattening([split_snake_case_name_to_words(function_name) for function_name in function_names])
+    
 
-# def get_top_functions_names_in_path(path, top_size=10):
-#     t = get_trees(path)
-#     nms = [f for f in flattening([[node.name.lower() for node in ast.walk(t) if isinstance(node, ast.Functi
-# onDef)] for t in t]) if not (f.startswith('__') and f.endswith('__'))]
-#     return collections.Counter(nms).most_common(top_size)
-
-
-# UNNESSESARY
-# def flatten(array):
-#     return [arr for arr in flattening(array)]
